backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
djangorestframework==3.16.1
dj-database-url==2.2.0
Django==6.0.3
cryptography==46.0.5
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.10; python_version < "3.14"
psycopg[binary]>=3.2,<3.3; python_version >= "3.14"
PyJWT==2.11.0
python-dotenv==1.2.2
PyYAML==6.0.3
whitenoise==6.7.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
